Apply and enforce format rules
Checkstyle plugin is used to enforce the format rules and hence the build will fail if a format rule is violated. 
The rules are specified in the `checkstyle.xml` file and mimic the ones defined in Quarkus upstream.
As part of this PR, I didn't format the ".properties" files or ".html" files.
Some rules:
- Avoid magic values/numbers
- Method name must match a reg expression
- Field declaration order
- Unused imports
- Imports in right order
- XML whitespaces vs tabs